requirements.txt
GitPython
Pillow